pyproject.toml
[project]
name = "lala"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Stepan", email = "stephan.savchenko@gmail.com" }
]
dependencies = [
    "httpx==0.25.2",
    "llama-cpp-python>=0.2.85",
    "uvicorn>=0.30.4",
    "litestar>=2.10.0",
    "pydantic>=2.9.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/lala"]
